modifica ad init.sh

+++ Dockers/scripts/init.sh
@@ -6,9 +6,12 @@
 CLEAN_INIT=false 
 
 #installazione e compilazione di sparqlify da src
-if [ "$CLEAN_INIT" = true ] || [ ! -d "/home/sparqlify" ];
+if [ "$CLEAN_INIT" = true ] && [ -d "/home/sparqlify" ];
 then 
     rm -fr /home/sparqlify
+    git clone https://github.com/SmartDataAnalytics/Sparqlify /home/sparqlify
+    cp -fr ./Main.java /home/sparqlify/sparqlify-cli/src/main/java/org/aksw/sparqlify/web/Main.java
+else
     git clone https://github.com/SmartDataAnalytics/Sparqlify /home/sparqlify
     cp -fr ./Main.java /home/sparqlify/sparqlify-cli/src/main/java/org/aksw/sparqlify/web/Main.java
 fi
